switch signal from int to float

+++ file_store/src/lora_witness_report.rs
@@ -12,7 +12,7 @@ pub struct LoraWitnessReport {
     pub data: Vec<u8>,
     pub timestamp: DateTime<Utc>,
     pub ts_res: u32,
-    pub signal: u32,
+    pub signal: f32,
     pub snr: f32,
     pub frequency: u64,
     pub datarate: DataRate,
